Update functions to support linux
Except `syncfromdotfiles`, I think this requires computer specific
rather than OS specific carve outs.

+++ .config/fish/functions/movemydockback.fish
@@ -1,4 +1,9 @@
 function movemydockback --description "Move my dock back from the external display to my laptop."
+	# Exit early if not MacOS
+    if test (uname) != "Darwin"
+        echo "The \`movemydockback\` function is only intended for use on MacOS."
+        return 1
+    end
 	defaults write com.apple.dock appswitcher-all-displays -bool true
 	killall Dock
 end
